We normal Japanese don't read 徴 as ち. It should NOT be translated.
It affects the entire branding of the project in Japanese, and
this part looks super awkward, looking totally bogus Japanese.
(There is no possible syntactical explanation why it looks awkward
in native Japanese, but I'm sure 99.99*% of Japanese would agree.)

Since Meilin part is left untranslated and it looks nice, the former
half should follow the same.

(Actually I don't follow the idea behind this "徴音" part in the
Chinese context, so it even looks to me that this can be removed
in Japanese documents. There is nothing like this in 夏語遙 or 心華.)